Creacion de metodos cargarpeliculas y cargarVideojuegos

+++ src/domain/Pelicula.java
@@ -11,7 +11,7 @@ public class Pelicula extends Producto{
 	private String director;
 	private int duracion;
 	
-	public Pelicula(long id,String titulo, String sinopsis, float precio, int rating, ArrayList<Review> comentarios,
+	public Pelicula(long id,String titulo, String sinopsis, double precio, int rating, ArrayList<Review> comentarios,
 			TipoPelicula tipo, GeneroPelicula genero, String director, int duracion, ImageIcon foto) {
 		super(id,titulo, sinopsis, precio, rating, comentarios, foto);
 		this.tipo = tipo;
